alteração permissão por username
alteração permissão por username

+++ bot.py
@@ -8,6 +8,9 @@ TOKEN = os.environ.get('TOKEN')
 # Diretório onde está o arquivo de log
 LOG_FILE_PATH = '/caminho/datalog.log'
 
+# Lista de usernames permitidos
+USERS_ALLOWED = ["raudeliunas", "alesinicio", "GustavoSobrinh0"]
+
 # Criar instância do bot
 bot = telebot.TeleBot(TOKEN)
 
@@ -16,18 +19,22 @@ bot = telebot.TeleBot(TOKEN)
 @bot.message_handler(commands=['enviarlog'])
 def enviar_log(message):
     try:
-        # Compactar o arquivo de log
-        with zipfile.ZipFile('log.zip', 'w') as zipf:
-            zipf.write(LOG_FILE_PATH, arcname=os.path.basename(LOG_FILE_PATH))
+        # Verificar se o remetente está na lista de usernames permitidos
+        if message.from_user.username in USERS_ALLOWED:
+            # Compactar o arquivo de log
+            with zipfile.ZipFile('log.zip', 'w') as zipf:
+                zipf.write(LOG_FILE_PATH, arcname=os.path.basename(LOG_FILE_PATH))
 
-        # Enviar o arquivo zip
-        with open('log.zip', 'rb') as zip_file:
-            bot.send_document(message.chat.id, zip_file)
+            # Enviar o arquivo zip
+            with open('log.zip', 'rb') as zip_file:
+                bot.send_document(message.chat.id, zip_file)
 
-        # Remover o arquivo zip temporário
-        os.remove('log.zip')
+            # Remover o arquivo zip temporário
+            os.remove('log.zip')
 
-        bot.reply_to(message, "Arquivo de log enviado com sucesso!")
+            bot.reply_to(message, "Arquivo de log enviado com sucesso!")
+        else:
+            bot.reply_to(message, "Você não tem permissão para usar este comando.")
     except Exception as e:
         bot.reply_to(message, f"Erro ao enviar arquivo de log: {e}")
